Update BUFR_TableD_en_10.csv
Last week our colleagues at ECMWF were kind enough to point out that during the preparation of our many inputs to the current Fast Track (FT 2023-I), we have inadvertently introduced the wrong replication factors in the proposed Table D sequence for MWI, 3-10-081.
The _correct_ fixed replication factor for the number of horns is 8. This is correct in the “Element name” column, but _wrong_ in the “Table references” column, where we refer to 7 horns.  The entry in the sequence should say 1 13 008,  Repeat 13 descriptors 8 times
 
The _correct_ fixed replication factors for the channel blocks are: 2, 2, 2, 8, 2, 4, 1, 5. This gives 26 channels.
We have passed you the following _wrong_ fixed replication factors: 1, 2, 3, 3, 2, 3, 3, 2. This would give 19 channels.
 
Changes are embedded with this commit (related to the following :  https://github.com/wmo-im/BUFR4/issues/136)

Thanks